third_party/blink/web_tests/svg/hixie/error/010-expected.txt

CONSOLE ERROR: Error: <svg> attribute viewBox: Trailing garbage, "0 0 1 1 1".
